Removed Java and Kotlin SourceSets
