+++ src/main/com/flaghacker/sttt/common/Bot.kt
