+++ src/main/com/flaghacker/sttt/common/Util.kt
